requirements.txt
requests
psycopg2-binary
pyodbc
sqlalchemy
python-dotenv
unixodbc